Add files via upload
Position salaries for the polynomial regression model. For models that form a relationship that is not appear linear (straight line), this provides a model and therefore a prediction of a dependent variable. In this example it is the position level of workers in a comapny and their salary. Evidently, the more successful you become, the faster the shmoney comes through.